+++ app/src/main/java/labs/zero_one/roundo/PreferenceMainActivity.kt
@@ -1,7 +1,11 @@
 package labs.zero_one.roundo
 
+import android.content.DialogInterface
+import android.content.Intent
 import android.content.SharedPreferences
+import android.os.Build
 import android.os.Bundle
+import android.provider.Settings
 import android.support.v7.app.AppCompatActivity
 import android.view.MenuItem
 import com.takisoft.fix.support.v7.preference.PreferenceFragmentCompat
@@ -46,6 +50,48 @@ class PreferenceMainActivity : AppCompatActivity() {
             ) {
                 findPreference(getString(R.string.pref_display_show_second_key)).icon =
                     requireContext().getDrawable(R.drawable.ic_pref_display_show_second_ee)
+            }
+            findPreference(getString(R.string.pref_other_background_key))
+                .setOnPreferenceClickListener {
+                    val romType = RomKit.getType()
+                    var negativeButtonListener: ((DialogInterface, Int) -> (Unit))? = null
+                    if (romType == RomKit.RomType.EMUI) {
+                        negativeButtonListener = { _, _ ->
+                            startActivity(RomKit.EMUI_BACKGROUND_MANAGER)
+                        }
+                    } else if (romType == RomKit.RomType.MIUI) {
+                        negativeButtonListener = { _, _ ->
+                            startActivity(RomKit.MIUI_BACKGROUND_MANAGER)
+                        }
+                    } else if (romType == RomKit.RomType.FLYME) {
+                        negativeButtonListener = { _, _ ->
+                            DialogKit.showDialog(
+                                requireContext(),
+                                R.string.pref_other_background_title,
+                                R.string.pref_other_background_settings_flyme,
+                                icon = requireContext()
+                                    .getDrawable(R.drawable.ic_pref_other_background)
+                            )
+                        }
+                    } else if (Build.VERSION.SDK_INT >= 23) {
+                        negativeButtonListener = { _, _ ->
+                            startActivity(Intent(Settings.ACTION_IGNORE_BATTERY_OPTIMIZATION_SETTINGS))
+                        }
+                    }
+                    DialogKit.showDialog(
+                        requireContext(),
+                        R.string.pref_other_background_title,
+                        R.string.pref_other_background_detail,
+                        negativeButtonTextId =
+                        if (negativeButtonListener != null)
+                            R.string.pref_other_background_settings
+                        else
+                            null,
+                        negativeButtonListener = negativeButtonListener,
+                        icon = requireContext().getDrawable(R.drawable.ic_pref_other_background),
+                        cancelable = false
+                    )
+                true
             }
 
             defaultSharedPreferences.registerOnSharedPreferenceChangeListener(this)